+++ src/util.cc
@@ -23,20 +23,20 @@
 #include <netlink/util.h>
 
 
-unsigned NL_NAMESPACE_NAME::getTime() {
+unsigned long long NL_NAMESPACE_NAME::getTime() {
 
     #ifdef OS_WIN32
 
         SYSTEMTIME now;
         GetSystemTime(&now);
-        unsigned milisec = now.wHour *3600*1000 + now.wMinute *60*1000 + now.wSecond *1000 + now.wMilliseconds;
+        unsigned long long milisec = now.wHour *3600*1000 + now.wMinute *60*1000 + now.wSecond *1000 + now.wMilliseconds;
         return(milisec);
 
     #else
 
         struct timeval now;
         gettimeofday(&now, NULL);
-        unsigned milisec = now.tv_sec * 1000 + now.tv_usec / 1000.0;
+        unsigned long long milisec = now.tv_sec * 1000 + now.tv_usec / 1000.0;
         return(milisec);
 
     #endif
